+++ src/components/NcBreadcrumb/NcBreadcrumb.vue
@@ -82,6 +82,7 @@ export default {
 		NcActions,
 		ChevronRight,
 	},
+	inheritAttrs: false,
 	props: {
 		/**
 		 * The main text content of the entry.